Create convert_x509_pkcs12.sh
The script prompts the user to enter the filenames of the certificate file and private key file in PEM format. The script then automatically creates the required variables for conversion to PKCS12 and x509.

The script creates a menu with three options: Convert x509 to PKCS12, Convert PKCS12 to x509, and Exit. Depending on the user's selection, the corresponding conversion script is executed.

Note that the script saves the PKCS12 file with an automatically generated filename in the same directory as the original certificate. The x509 certificate is saved with an automatically generated filename in the same directory as the original certificate, but with the suffix "_converted.crt".